requirements.txt
Django>=5.0
Pillow>=10.2
django-modeltranslation>=0.18
psycopg2-binary>=2.9
django-cleanup>=8.0
python-decouple>=3.8
gunicorn>=21.2
